+++ deep_hedging_v0/hedging/black_scholes_model.py
@@ -12,5 +12,5 @@ def black_scholes_pricer(K, T, S, r, sigma, call_flag):
     d1, d2 = black_scholes_d1_d2(K, T, S, r, sigma)
     term1 = call_flag * S * norm.cdf(call_flag * d1)
     term2 = call_flag * K * np.exp(-r * T) * norm.cdf(call_flag * d2)
-    option_prices = term1 - term2
+    option_prices = (term1 - term2).astype(np.float32, copy=False)
     return option_prices
